chore: move more components around
