+++ packages/api-reference/package.json
@@ -83,10 +83,8 @@
     "rehype-highlight": "^7.0.0",
     "rehype-raw": "^7.0.0",
     "rehype-sanitize": "^6.0.0",
-    "rehype-slug-custom-id": "^1.1.0",
     "rehype-stringify": "^10.0.0",
     "remark-gfm": "^4.0.0",
-    "remark-html": "^16.0.1",
     "remark-parse": "^11.0.0",
     "remark-rehype": "^11.1.0",
     "remark-stringify": "^11.0.0",
@@ -105,6 +103,7 @@
     "@types/js-yaml": "^4.0.6",
     "@vitejs/plugin-vue": "^4.4.0",
     "@vitest/coverage-v8": "^0.34.4",
+    "eslint": "^8.56.0",
     "http-server": "^14.1.1",
     "postcss": "^8.4.31",
     "react": "^18.2.0",
