+++ packages/api-reference/src/components/DarkModeToggle/index.ts
@@ -1,1 +1,2 @@
 export { default as DarkModeToggle } from './DarkModeToggle.vue'
+export { default as DarkModeIconToggle } from './DarkModeIconToggle.vue'
