+++ packages/api-reference/src/index.ts
@@ -6,10 +6,10 @@ export { default as ApiReference } from './components/ApiReference.vue'
 export { default as ApiReferenceLayout } from './components/ApiReferenceLayout.vue'
 export { default as ApiReferenceBase } from './components/ApiReferenceBase.vue'
 export { default as RenderedReference } from './components/Content/Content.vue'
-export { default as DarkModeToggle } from './components/DarkModeToggle.vue'
 export { default as SearchModal } from './components/SearchModal.vue'
 export { default as SearchButton } from './components/SearchButton.vue'
 
+export * from './components/DarkModeToggle'
 export * from './components/Sidebar'
 export * from './components/Card'
